+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 - Diff control: `unnest_records` setting to "flatten" nested hashes. A nested value appears with its ancestor keys on the same line.
 - Diff control: `sort_nested` setting defaults to true. If true, unnested hashes will be sorted. If false the order will depend on the calling code.
 ### Changed
+- Input control: Add `:alphanumeric` pattern and set the `title` by default when a built-in pattern is used.
 ### Fixed
 
 ## [0.4.4] - 2020-07-04